Move `MarkdownType` and `MarkdownRange` to commonTypes.ts
