+++ src/__tests__/webParser.test.tsx
@@ -1,7 +1,6 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import {expect} from '@jest/globals';
 import {parseRangesToHTMLNodes} from '../web/utils/parserUtils';
-import type {MarkdownRange} from '../web/utils/parserUtils';
 
 require('../../parser/react-native-live-markdown-parser.js');
 
@@ -16,8 +15,7 @@ const toBeParsedAsHTML = function (actual: string, expectedHTML: string) {
     throw new Error('Actual value must be a string');
   }
   let expected = expectedHTML;
-  const ranges = global.parseExpensiMarkToRanges(actual);
-  const markdownRanges = ranges as MarkdownRange[];
+  const markdownRanges = global.parseExpensiMarkToRanges(actual);
 
   const actualDOM = parseRangesToHTMLNodes(actual, markdownRanges, {}, true).dom;
   const actualHTML = actualDOM.innerHTML;
